Save and Update of inmueble

+++ Models/Inmueble.cs
@@ -27,7 +27,7 @@ namespace inmoCabreraNet.Models
         public int inm_uso { get; set; }
         public string usoNombre => inm_uso > 0 ? ((enUsos)inm_uso).ToString() : "";
         public int inm_ambientes { get; set; }
-        public decimal inm_precio { get; set; }
+        public double inm_precio { get; set; }
         public bool inm_disponible { get; set; }
         public string disponibleNombre => inm_disponible ? "Sí" : "No";
         public int inm_pro_id { get; set; }
